Move IProductService to ServiceInterfaces folder and adjusted namespaces
